python/RAG-MODULE/image-processing/requirements.txt
PyMuPDF==1.23.8
pillow==10.0.1
pytesseract==0.3.10
uvicorn==0.24.0
python-multipart==0.0.6
